added tests for intersect_flex frames
