+++ tests/tools/test_intersect.py
@@ -31,43 +31,6 @@ TESTCASES = [  # startdates, freq, expected_startdate
 ]
 
 
-@pytest.mark.parametrize("types", ["series", "df"])
-@pytest.mark.parametrize("ignore_tz", [True, False])
-def test_frames_ignore_tz(types: str, ignore_tz: bool):
-    idx_a = pd.date_range(
-        "2020", "2022", freq="MS", inclusive="left", tz="Europe/Berlin"
-    )
-    a = pd.Series(range(0, 24), idx_a)
-    idx_b = pd.date_range("2020-02", "2021-09", freq="MS", inclusive="left")
-    b = pd.Series(range(0, 19), idx_b)
-
-    exp_idx_a = pd.date_range(
-        "2020-02", "2021-09", freq="MS", inclusive="left", tz="Europe/Berlin"
-    )
-    exp_idx_b = idx_b
-    exp_a = pd.Series(range(1, 20), exp_idx_a)
-    exp_b = pd.Series(range(0, 19), exp_idx_b)
-
-    if types == "series":
-        if not ignore_tz:
-            with pytest.raises(ValueError):
-                _ = tools.intersect.frames(a, b, ignore_tz=ignore_tz)
-            return
-        result_a, result_b = tools.intersect.frames(a, b, ignore_tz=ignore_tz)
-        testing.assert_series_equal(result_a, exp_a)
-        testing.assert_series_equal(result_b, exp_b)
-    else:
-        a, b = pd.DataFrame({"col_a": a}), pd.DataFrame({"col_b": b})
-        if not ignore_tz:
-            with pytest.raises(ValueError):
-                _ = tools.intersect.frames(a, b, ignore_tz=ignore_tz)
-            return
-        exp_a, exp_b = pd.DataFrame({"col_a": exp_a}), pd.DataFrame({"col_b": exp_b})
-        result_a, result_b = tools.intersect.frames(a, b, ignore_tz=ignore_tz)
-        testing.assert_frame_equal(result_a, exp_a)
-        testing.assert_frame_equal(result_b, exp_b)
-
-
 def get_idx(
     startdate: str, starttime: str, tz: str, freq: str, enddate: str = COMMON_END
 ) -> pd.DatetimeIndex:
